more testing and code clean up
