+++ app/src/main/java/com/example/clicker/domain/GetFollowedLiveStreamsUseCase.kt
@@ -3,12 +3,11 @@ package com.example.clicker.domain
 import com.example.clicker.network.domain.TwitchRepo
 import com.example.clicker.network.models.twitchRepo.FollowedLiveStreams
 import com.example.clicker.network.models.twitchRepo.toStreamInfo
-import com.example.clicker.presentation.home.StreamInfo
 import com.example.clicker.util.Response
 import javax.inject.Inject
 
-class GetFollowedLiveStreamsUseCase @Inject constructor(
-    private val twitchRepoImpl: TwitchRepo
+class GetFollowedLiveStreamsUseCase constructor(
+
 ) {
     suspend operator fun invoke(authorizationToken: String, clientId: String, userId: String) {
 //        val items =twitchRepoImpl.getFollowedLiveStreams(
@@ -22,11 +21,4 @@ class GetFollowedLiveStreamsUseCase @Inject constructor(
 //        }
     }
 
-    private fun followedLiveStreamToStreamInfo(stream: Response<FollowedLiveStreams>): Response<List<StreamInfo>> {
-        return when (stream) {
-            is Response.Loading -> Response.Loading
-            is Response.Success -> Response.Success(stream.data.data.map { it.toStreamInfo() })
-            is Response.Failure -> Response.Failure(stream.e)
-        }
-    }
 }
