+++ app/src/test/java/com/example/clicker/presentation/HomeViewModelTest.kt
@@ -2,13 +2,11 @@ package com.example.clicker.presentation
 
 import com.example.clicker.data.TokenDataStore
 import com.example.clicker.domain.TwitchDataStore
-import com.example.clicker.network.clients.GetModChannels
 import com.example.clicker.network.domain.TwitchAuthentication
 import com.example.clicker.network.domain.TwitchRepo
 import com.example.clicker.network.models.twitchAuthentication.ValidatedUser
 import com.example.clicker.network.models.twitchRepo.StreamData
 import com.example.clicker.presentation.home.HomeViewModel
-import com.example.clicker.presentation.home.StreamInfo
 import com.example.clicker.presentation.util.FakeAuthentication
 import com.example.clicker.presentation.util.FakeTokenDataStore
 import com.example.clicker.presentation.util.FakeTwitchImplRepo
@@ -141,213 +139,6 @@ class HomeViewModelTest {
         Assert.assertEquals(expectedValue.javaClass, actualValue.javaClass)
 
     }
-//
-//    @Test
-//    fun testing_get_live_channels_fail() = runTest {}
-//
-//
-//
-//    @Test
-//    fun validateOAuthToken_generic_error() = runTest {
-//        /**GIVEN*/
-//        val dispatcher = StandardTestDispatcher(testScheduler)
-//        val fakeAuthentication =FakeAuthentication()
-//        val fakeReturnType =NetworkNewUserResponse.Failure(Exception("Failed to authenticate"))
-//        fakeAuthentication.setValidateTokenReturnType(fakeReturnType)
-//        val homeViewModel:HomeViewModel = HomeViewModel(
-//            ioDispatcher =dispatcher,
-//            authentication = fakeAuthentication,
-//            twitchRepoImpl = FakeTwitchImplRepo(),
-//            tokenDataStore = FakeTokenDataStore(
-//                userIsNewUser = false
-//            )
-//        )
-//        /**WHEN*/
-//        delay(1000)
-//        val actualUIState = homeViewModel.state.value.streamersListLoading
-//        val expectedUIState= fakeReturnType
-//
-//
-//        /**THEN*/
-//
-//        Assert.assertEquals(expectedUIState, actualUIState)
-//
-//    }
-//    @Test
-//    fun validateOAuthToken_401Authentication_error() = runTest {
-//        /**GIVEN*/
-//        val dispatcher = StandardTestDispatcher(testScheduler)
-//        val fakeAuthentication =FakeAuthentication()
-//        val fakeReturnType =NetworkNewUserResponse.Auth401Failure(Exception("Token Expired"))
-//        fakeAuthentication.setValidateTokenReturnType(fakeReturnType)
-//        val homeViewModel:HomeViewModel = HomeViewModel(
-//            ioDispatcher =dispatcher,
-//            authentication = fakeAuthentication,
-//            twitchRepoImpl = FakeTwitchImplRepo(),
-//            tokenDataStore = FakeTokenDataStore(
-//                userIsNewUser = false
-//            )
-//        )
-//        /**WHEN*/
-//        delay(1000)
-//        val actualUIState = homeViewModel.state.value.streamersListLoading
-//        val expectedUIState= fakeReturnType
-//
-//
-//        /**THEN*/
-//
-//        Assert.assertEquals(expectedUIState, actualUIState)
-//
-//    }
-//
-//    @Test
-//    fun validateOAuthToken_NetworkFailure_error() = runTest {
-//        /**GIVEN*/
-//        val dispatcher = StandardTestDispatcher(testScheduler)
-//        val fakeAuthentication =FakeAuthentication()
-//        val fakeReturnType =NetworkNewUserResponse.NetworkFailure(Exception("Network error"))
-//        fakeAuthentication.setValidateTokenReturnType(fakeReturnType)
-//        val homeViewModel:HomeViewModel = HomeViewModel(
-//            ioDispatcher =dispatcher,
-//            authentication = fakeAuthentication,
-//            twitchRepoImpl = FakeTwitchImplRepo(),
-//            tokenDataStore = FakeTokenDataStore(
-//                userIsNewUser = false
-//            )
-//        )
-//        /**WHEN*/
-//        delay(1000)
-//        val actualUIState = homeViewModel.state.value.streamersListLoading
-//        val expectedUIState= fakeReturnType
-//
-//
-//        /**THEN*/
-//
-//        Assert.assertEquals(expectedUIState, actualUIState)
-//
-//    }
-//
-//    @Test
-//    fun beginLogout_success() = runTest {
-//        /**GIVEN*/
-//        val dispatcher = StandardTestDispatcher(testScheduler)
-//        val fakeAuthentication =FakeAuthentication()
-//        val fakeReturnType =NetworkAuthResponse.Success(false)
-//        fakeAuthentication.setLogoutReturnType(fakeReturnType)
-//
-//        val homeViewModel:HomeViewModel = HomeViewModel(
-//            ioDispatcher =dispatcher,
-//            authentication = fakeAuthentication,
-//            twitchRepoImpl = FakeTwitchImplRepo(),
-//            tokenDataStore = FakeTokenDataStore(
-//                userIsNewUser = true
-//            )
-//        )
-//        /**WHEN*/
-//        delay(1000)
-//        homeViewModel.beginLogout("fakeClientId","fakeoAuthToken")
-//        delay(1000)
-//        val actualUIState= homeViewModel.state.value.userIsLoggedIn
-//        val expectedUIState =fakeReturnType
-//
-//
-//
-//        /**THEN*/
-//
-//        Assert.assertEquals(expectedUIState, actualUIState)
-//
-//    }
-//
-//    @Test
-//    fun beginLogout_NetworkAuthResponse_Failure() = runTest {
-//        /**GIVEN*/
-//        val dispatcher = StandardTestDispatcher(testScheduler)
-//        val fakeAuthentication =FakeAuthentication()
-//        val fakeReturnType =NetworkAuthResponse.Failure(Exception("fake exception"))
-//        fakeAuthentication.setLogoutReturnType(fakeReturnType)
-//
-//        val homeViewModel:HomeViewModel = HomeViewModel(
-//            ioDispatcher =dispatcher,
-//            authentication = fakeAuthentication,
-//            twitchRepoImpl = FakeTwitchImplRepo(),
-//            tokenDataStore = FakeTokenDataStore(
-//                userIsNewUser = true
-//            )
-//        )
-//        /**WHEN*/
-//        delay(1000)
-//        homeViewModel.beginLogout("fakeClientId","fakeoAuthToken")
-//        delay(1000)
-//        val actualUIState= homeViewModel.state.value.userIsLoggedIn
-//        val expectedUIState =fakeReturnType
-//
-//
-//
-//        /**THEN*/
-//
-//        Assert.assertEquals(expectedUIState, actualUIState)
-//
-//    }
-//
-//    @Test
-//    fun beginLogout_NetworkAuthResponse_NetworkFailure() = runTest {
-//        /**GIVEN*/
-//        val dispatcher = StandardTestDispatcher(testScheduler)
-//        val fakeAuthentication =FakeAuthentication()
-//        val fakeReturnType =NetworkAuthResponse.NetworkFailure(Exception("fake exception"))
-//        fakeAuthentication.setLogoutReturnType(fakeReturnType)
-//
-//        val homeViewModel:HomeViewModel = HomeViewModel(
-//            ioDispatcher =dispatcher,
-//            authentication = fakeAuthentication,
-//            twitchRepoImpl = FakeTwitchImplRepo(),
-//            tokenDataStore = FakeTokenDataStore(
-//                userIsNewUser = true
-//            )
-//        )
-//        /**WHEN*/
-//        delay(1000)
-//        homeViewModel.beginLogout("fakeClientId","fakeoAuthToken")
-//        delay(1000)
-//        val actualUIState= homeViewModel.state.value.userIsLoggedIn
-//        val expectedUIState =fakeReturnType
-//
-//
-//
-//        /**THEN*/
-//
-//        Assert.assertEquals(expectedUIState, actualUIState)
-//
-//    }
-//    @Test
-//    fun beginLogout_NetworkAuthResponse_Auth401Failure() = runTest {
-//        /**GIVEN*/
-//        val dispatcher = StandardTestDispatcher(testScheduler)
-//        val fakeAuthentication =FakeAuthentication()
-//        val fakeReturnType =NetworkAuthResponse.Auth401Failure(Exception("fake exception"))
-//        fakeAuthentication.setLogoutReturnType(fakeReturnType)
-//
-//        val homeViewModel:HomeViewModel = HomeViewModel(
-//            ioDispatcher =dispatcher,
-//            authentication = fakeAuthentication,
-//            twitchRepoImpl = FakeTwitchImplRepo(),
-//            tokenDataStore = FakeTokenDataStore(
-//                userIsNewUser = true
-//            )
-//        )
-//        /**WHEN*/
-//        delay(1000)
-//        homeViewModel.beginLogout("fakeClientId","fakeoAuthToken")
-//        delay(1000)
-//        val actualUIState= homeViewModel.state.value.userIsLoggedIn
-//        val expectedUIState =fakeReturnType
-//
-//
-//
-//        /**THEN*/
-//
-//        Assert.assertEquals(expectedUIState, actualUIState)
-//
-//    }
+
 }
